% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{bond}
\alias{bond}
\title{Create \code{bond} Object}
\usage{
bond(maturity, coupon, dated_date, conv = "30/360", freq = 2,
  redemption = 100, id = NULL, cfs = NULL, call_date = NULL,
  call_px = NULL)
}
\arguments{
\item{maturity}{Maturity/Workout date}

\item{coupon}{Annnual coupon amount}

\item{dated_date}{Dated date, usually the issue date and the day from which
interest accrues.}

\item{conv}{Daycount convention (one of \code{"30/360"}, \code{"Act/Act"},
or \code{"Act/365"})}

\item{freq}{Coupon frequency (number of periods per year)}

\item{redemption}{Redemption value or par amount}

\item{id}{CUSIP or other id}

\item{cfs}{List of cashflow tbl_dfs (id, cf_date, cf_amount)}
}
\value{
\code{bond} object with cashflows
}
\description{
Given basic bond details like maturity date, call date and price (if callable),
coupon, frequency, settle, daycount convention, and id (e.g. cusip),
this will create an object of type \code{bond}. The constructor can take a
\code{cfs} object, which, if passed, should be a list of \code{tbl_df}s, where each
one has a name that means something. e.g. \code{cfs$maturity}, \code{cfs$call},
\code{cfs$put} etc. \code{bond} is the basic object in this package. It is used
to calculate prices, yields, risk metrics, and do any other kind of fixed
income analysis.
}
\examples{
bond("2045-02-15", 2.5, "2015-02-17", "Act/Act", 2, 100, id = "ABCD")
bond("2045-02-15", 2.5, "2015-02-17", "Act/Act", 2, 100, id = "ABCD",
     call_date = "2025-02-15", call_px = 101)
}

