% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{calc_dirty_px}
\alias{calc_dirty_px}
\alias{calc_dirty_px.bond}
\alias{calc_dirty_px.default}
\title{Calculate Dirty Price of a Bond}
\usage{
calc_dirty_px(x, ...)

\method{calc_dirty_px}{default}(maturity, settle, coupon, yield,
  conv = "30/360", freq = 2)

\method{calc_dirty_px}{bond}(b, settle, yield)
}
\arguments{
\item{maturity}{Maturity/Workout date}

\item{settle}{Settlement Date}

\item{coupon}{Annnual coupon amount}

\item{yield}{Yield in percent (e.g. 5.0)}

\item{conv}{Daycount convention (one of \code{"30/360"}, \code{"Act/Act"},
or \code{"Act/365"})}

\item{freq}{Coupon frequency (number of periods per year)}

\item{b}{A \code{bond} object}
}
\value{
Numeric value of the clean price and corresponding worst workout date
if there are multiple sets of cashflows in the bond
}
\description{
Given a \code{bond} or the necessary parameters to construct one, this
will calculate the price of the \code{bond} including accrued interest. This
price is used in yield calculations and other functions that require the
"full" or "dirty" price.
}
\examples{
calc_dirty_px("2045-02-15", "2015-02-15", 4, 3, "Act/Act", 2)
b <- bond("2045-02-15", 2.5, "2015-02-17", "Act/Act", 2, 100, id = "ABCD")
calc_dirty_px(b, "2015-05-15", 3)
}
\seealso{
Other pricing functions: \code{\link{acc_int}};
  \code{\link{calc_clean_px}},
  \code{\link{calc_clean_px.bond}},
  \code{\link{calc_clean_px.default}};
  \code{\link{calc_risk}}, \code{\link{calc_risk.bond}},
  \code{\link{calc_risk.default}};
  \code{\link{calc_yield}}, \code{\link{calc_yield.bond}},
  \code{\link{calc_yield.default}};
  \code{\link{discount_cfs_single}};
  \code{\link{risk_cfs_single}};
  \code{\link{yield_cfs_single}}; \code{\link{ytm}}
}

