% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{calc_risk}
\alias{calc_dv01}
\alias{calc_dv01.bond}
\alias{calc_dv01.default}
\alias{calc_risk}
\alias{calc_risk.bond}
\alias{calc_risk.default}
\title{Calculate the Risks of a Bond}
\usage{
calc_risk(x, ...)

\method{calc_risk}{default}(maturity, settle, coupon, price_yield,
  input_type = "Y", returnCFs = FALSE, conv = "30/360", freq = 2)

\method{calc_risk}{bond}(b, settle, price_yield, input_type = "Y",
  returnCFs = FALSE)

calc_dv01(x, ...)

\method{calc_dv01}{default}(maturity, settle, coupon, price_yield,
  input_type = "Y", conv = "30/360", freq = 2)

\method{calc_dv01}{bond}(b, settle, price_yield, input_type = "Y")
}
\arguments{
\item{maturity}{Maturity/Workout date}

\item{settle}{Settlement Date}

\item{coupon}{Annnual coupon amount}

\item{price_yield}{The clean price of the bond or the yield to worst}

\item{input_type}{\code{"Y"} for Yield to Worst (default) or \code{"P"}
for clean price}

\item{returnCFs}{Boolean flag indicating whether or not to return the
calculated set of cashflows and risk metrics}

\item{conv}{Daycount convention (one of \code{"30/360"}, \code{"Act/Act"},
or \code{"Act/365"})}

\item{freq}{Coupon frequency (number of periods per year)}

\item{b}{A \code{bond} object}
}
\value{
Numeric value of the DV01 and other risks and corresponding worst
workout date and workout type if there are multiple sets of cashflows
in the bond or a full set of calculated cashflows and risk metrics
as a \code{tbl_df}
}
\description{
Given a \code{bond} or the necessary parameters to construct one, this
will calculate the DV01 and other risks of the \code{bond} for the worst
workout date. The workout dates used are the based on the \code{cfs}
object's last cashflow dates in the \code{bond}. This function uses
\code{discount_cfs_single} to do the actual calculations.
}
\examples{
calc_risk("2045-02-15", "2015-02-15", 4, 101, "P", TRUE, "Act/Act", 2)
b <- bond("2045-02-15", 2.5, "2015-02-17", "Act/Act", 2, 100, id = "ABCD")
calc_risk(b, "2015-05-15", 101, "P", TRUE)
calc_dv01(b, "2015-05-15", 3, "Y")
}
\seealso{
Other pricing functions: \code{\link{acc_int}};
  \code{\link{calc_clean_px}},
  \code{\link{calc_clean_px.bond}},
  \code{\link{calc_clean_px.default}};
  \code{\link{calc_dirty_px}},
  \code{\link{calc_dirty_px.bond}},
  \code{\link{calc_dirty_px.default}};
  \code{\link{calc_yield}}, \code{\link{calc_yield.bond}},
  \code{\link{calc_yield.default}}; \code{\link{calc_ytm}};
  \code{\link{discount_cfs_single}};
  \code{\link{risk_cfs_single}};
  \code{\link{yield_cfs_single}}
}

