% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{calc_yield}
\alias{calc_yield}
\alias{calc_yield.bond}
\alias{calc_yield.default}
\title{Calculate Yield to Worst of a Bond}
\usage{
calc_yield(x, ...)

\method{calc_yield}{default}(maturity, settle, coupon, clean_px,
  conv = "30/360", freq = 2)

\method{calc_yield}{bond}(b, settle, clean_px)
}
\arguments{
\item{maturity}{Maturity/Workout date}

\item{settle}{Settlement Date}

\item{coupon}{Annnual coupon amount}

\item{clean_px}{The clean price of the bond}

\item{conv}{Daycount convention (one of \code{"30/360"}, \code{"Act/Act"},
or \code{"Act/365"})}

\item{freq}{Coupon frequency (number of periods per year)}

\item{b}{A \code{bond} object}
}
\value{
Numeric value of the yield to worst and corresponding worst workout date
if there are multiple sets of cashflows in the bond
}
\description{
Given a \code{bond} or the necessary parameters to construct one, this
will calculate the yield to worst of the \code{bond}. The workout dates
used are the based on the \code{cfs} object's last cashflow dates in
the \code{bond}. This function uses \code{yield_cfs_single}, which in
turn uses the imported function \code{NMOF::ytm} to do the actual
calculations.
}
\examples{
calc_yield("2045-02-15", "2015-02-15", 4, 101, "Act/Act", 2)
b <- bond("2045-02-15", 2.5, "2015-02-17", "Act/Act", 2, 100, id = "ABCD")
calc_yield(b, "2015-05-15", 101)
}
\seealso{
Other pricing functions: \code{\link{acc_int}};
  \code{\link{calc_clean_px}},
  \code{\link{calc_clean_px.bond}},
  \code{\link{calc_clean_px.default}};
  \code{\link{calc_dirty_px}},
  \code{\link{calc_dirty_px.bond}},
  \code{\link{calc_dirty_px.default}};
  \code{\link{calc_risk}}, \code{\link{calc_risk.bond}},
  \code{\link{calc_risk.default}};
  \code{\link{discount_cfs_single}};
  \code{\link{risk_cfs_single}};
  \code{\link{yield_cfs_single}}; \code{\link{ytm}}
}

