% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{create_cf_amounts}
\alias{create_cf_amounts}
\title{Create vector of cashflow amounts}
\usage{
create_cf_amounts(num, coupon, freq, redemption = 100)
}
\arguments{
\item{num}{Number of cashflows}

\item{coupon}{Constant cashflow amount}

\item{freq}{Cashflow frequency (number of periods per year)}

\item{redemption}{Redemption value or par amount}
}
\value{
Vector of cashflow amounts of length \code{num} whose last value is
   \eqn{redemption + \frac{coupon}{freq}}{redemption + coupon / freq}.
}
\description{
Create vector of cashflow amounts
}
\examples{
create_cf_amounts(20, 5, 2, 100)
}
\seealso{
Other cashflows: \code{\link{create_cashflows}}
}

