% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dbQuery}
\alias{dbQuery}
\title{Query Database and Return Results}
\usage{
dbQuery(query, stringsAsFactors = FALSE, server = c_server,
  database = c_database, uid = c_uid, pwd = c_pwd)
}
\arguments{
\item{query}{Character string database query in MS SQL SERVER T-SQL format.}

\item{stringsAsFactors}{Boolean value as in base R}

\item{server}{Server name as a character string}

\item{database}{Database name as a character string}

\item{uid}{User ID as a character string}

\item{pwd}{Password as a character string}
}
\value{
\code{tbl_df} or vector containing the results of the query
}
\description{
Wrapper for RODBC odbcDriverConnect with default values for cetain
parameters. The connection defaults to the production server and
the Muni DB. The result is returned as a \code{tbl_df} with strings
represented as characters instead of factors, unless it is only one
column, in which case it is returned as a vector.
}
\examples{
qry <- paste("R_LoadMuniDeskPositions '",
           as.character(c_close_date),"','", as.character(c_as_of_datetime),"','",
           paste(c_exempt_accounts, c_taxable_accounts,
           c_tsy_accounts,sep = ",",collapse=","),"'")
dbQuery(qry)
}

