% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date_utils.R
\name{dc}
\alias{dc}
\alias{dc_30_360}
\alias{dc_act_act}
\title{Day Count}
\usage{
dc(date1, date2, conv = "30/360", eom = FALSE)
}
\arguments{
\item{date1}{starting date to begin counting}

\item{date2}{ending date to stop counting}

\item{conv}{the day count convention can be either "30/360" (default), "Act/365",
or "Act/Act"}

\item{eom}{option to assign additional days to February for the 30/360 day count
convention}
}
\value{
number of days between \code{date1} and \code{date2}
}
\description{
Counts the number of days between \code{date1} and \code{date2} via the provided day
count convention. For the 30/360 day count convention, it is assumed that there are
30 days in a month. If the bond is EOM, which is available only for the 30/360 day count
convention, in a regular year the 3 days between February 28 and March 1 is assigned
to February, otherwise to March. In a leap year, only 2 days is assigned to February
for EOM. If the \code{eom} parameter is set to \code{TRUE}, then two additional rules
will apply to EOM February dates to achieve the aforementioned behavior.
}
\examples{
d1 <- as.Date(c("2004-02-25","2003-02-28")
d2 <- as.Date(c("2015-01-01","2015-08-08")
dc(d1, d2, conv = "30/360")
dc(d1, d2, conv = "30/360", eom = TRUE)
}

