% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependent_constants.R
\name{get_close_date}
\alias{get_close_date}
\title{Get Previous Close Date}
\usage{
get_close_date(date = Sys.Date())
}
\arguments{
\item{date}{The date to use}
}
\value{
The previous business date using the SIFMA Calednar
}
\description{
Uses the Architect DB to get the business date before
the argument's business date.
}
\examples{
get_close_date()
}

