% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependent_constants.R
\name{get_eod}
\alias{get_eod}
\title{Get End of Day}
\usage{
get_eod(date = Sys.time())
}
\arguments{
\item{date}{The date to use}
}
\value{
POSIX variable for the date with hours, minutes,
and seconds equal to 23:59:59
}
\description{
Used to get a date with hours and minutes for the last
second of the day. This guarantees that all trades on a given
date will be captured.
}
\examples{
get_eod()
}

