% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{load_config}
\alias{load_config}
\title{Load Configurations from a File}
\usage{
load_config(filename)
}
\arguments{
\item{filename}{Path and filename to be loaded}
}
\value{
Nothing. Assignment occurs directly
}
\description{
Given a .txt filename, this will parse out key value pairs separated
by "=" and assign the value to the key in the environment which called
the function. The file should end in a new line character to avoid any
warnings.
}

