% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rights.R
\name{mdp_learnable}
\alias{mdp_learnable}
\title{Keep only rows eligible for derivative learning}
\usage{
mdp_learnable(x)
}
\arguments{
\item{x}{A data frame or lazy table with learning_eligible.}
}
\value{
The filtered input.
}
\description{
Keep only rows eligible for derivative learning
}
