% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{mdp_save_view}
\alias{mdp_save_view}
\title{Save a query as a view in your sandbox}
\usage{
mdp_save_view(x, name, con)
}
\arguments{
\item{x}{Lazy query retaining rights columns.}

\item{name}{Lowercase SQL identifier.}

\item{con}{PostgreSQL connection.}
}
\value{
The database execution result, invisibly.
}
\description{
Save a query as a view in your sandbox
}
