% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mdp_tbl}
\alias{mdp_tbl}
\title{Open a lazy warehouse relation}
\usage{
mdp_tbl(con, name, schema = "marts")
}
\arguments{
\item{con}{A DBI connection.}

\item{name}{Relation name.}

\item{schema}{Warehouse schema.}
}
\value{
A lazy dbplyr table.
}
\description{
Open a lazy warehouse relation
}
