\name{mdpr-package}
\alias{mdpr-package}
\alias{mdpr}
\docType{package}
\title{
\packageTitle{mdpr}
}
\description{
\packageDescription{mdpr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mdpr}
\packageIndices{mdpr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mdpr}

Maintainer: \packageMaintainer{mdpr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
