% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdpr.R
\docType{package}
\name{mdpr}
\alias{mdpr}
\alias{mdpr-package}
\title{mdpr: Municipal Data and Pricing}
\description{
mdpr will be used by traders of fixed income instruments, mainly municipal and US
government bonds. Functionality will include pricing (clean and dirty price, yield to
worst, and assorted other functions), risk analysis (dv01, convexity, etc.), and relative
value metrics.
}

