% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{risk_cfs_single}
\alias{risk_cfs_single}
\title{DV01 and Other Risks for a Single Set of Cashflows}
\usage{
risk_cfs_single(cfs, settle, yield, conv = "30/360", freq = 2,
  returnCFs = TRUE)
}
\arguments{
\item{cfs}{Cashflows \code{tbl_df}}

\item{settle}{Settlement Date}

\item{yield}{Yield or IRR used to discount the cashflows}

\item{conv}{Daycount convention (one of \code{"30/360"}, \code{"Act/Act"},
or \code{"Act/365"})}

\item{freq}{Coupon frequency (number of periods per year)}

\item{returnCFs}{Boolean flag indicating whether or not to return the
calculated set of cashflows and risk metrics}
}
\value{
Numeric value of the DV01 and other risks for the given cashflows
and yield or the full calculated set of risks by cashflow
}
\description{
Used as the backbone for all DV01 and risk calculations. This function uses
\code{discount_cfs_single} to calculate the important risk metrics.
}
\examples{
cfs <- create_cashflows("2045-02-15", 2.5, "2015-05-07", "Act/Act", 2, 100, "ABCD")
risk_cfs_single(cfs, "2015-05-07", 4, "30/360", 2, TRUE)
}
\seealso{
Other pricing functions: \code{\link{acc_int}};
  \code{\link{calc_clean_px}},
  \code{\link{calc_clean_px.bond}},
  \code{\link{calc_clean_px.default}};
  \code{\link{calc_dirty_px}},
  \code{\link{calc_dirty_px.bond}},
  \code{\link{calc_dirty_px.default}};
  \code{\link{calc_risk}}, \code{\link{calc_risk.bond}},
  \code{\link{calc_risk.default}};
  \code{\link{calc_yield}}, \code{\link{calc_yield.bond}},
  \code{\link{calc_yield.default}};
  \code{\link{discount_cfs_single}};
  \code{\link{yield_cfs_single}}; \code{\link{ytm}}
}

