% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{round_any_vec}
\alias{round_any_vec}
\title{Vectorized function for rounding to multiple of any number}
\usage{
round_any_vec(x, accuracy, f = round)
}
\arguments{
\item{x}{vector of numbers to round}

\item{accuracy}{vector of multiples to round the vector \code{x}}
}
\value{
vector with \code{x} rounded to the nearest \code{accuracy}
}
\description{
Vectorized version of \code{plyr::round_any}, which is a function that rounds
to multiple of any number.
}
\examples{
round_any_vec(c(52,199,14),10)
}

