% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{sqt}
\alias{sqt}
\title{Wrap Text in Apostrophes}
\usage{
sqt(text)
}
\arguments{
\item{text}{Text to be wrapped in apostrophes}
}
\value{
The text wrapped in apostrophes
}
\description{
Returns its argument wrapped in apostrophes. This can be used
when passing dates to SQL SERVER or for other similar
applications.
}
\examples{
sqt("2015-05-01")
}

