% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date_utils.R
\name{try_parse_date}
\alias{try_parse_date}
\title{Parse Date If Necessary}
\usage{
try_parse_date(d)
}
\arguments{
\item{d}{vector of elements that may or may not be type Date}
}
\value{
Date type vector with NA for a particular element in \code{d} if it were
not parsable
}
\description{
Since atomic objects can only have one type, it is unnecessary to check the type of
each element in the vector. This function checks the type on the vector level to
determine whether it needs to operate \code{parse_date_time} for \code{d} on a
vectorized basis.
}
\examples{
try_parse_date(as.Date(c("2015-01-01","2015-08-08")))
try_parse_date(c("2015-01-01","blah"))
}

