% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{yield_cfs_single}
\alias{yield_cfs_single}
\title{Yield or IRR for a Single Set of Cashflows}
\usage{
yield_cfs_single(cfs, settle, coupon, conv = "30/360", freq = 2, clean_px)
}
\arguments{
\item{cfs}{Cashflows \code{tbl_df}}

\item{settle}{Settlement Date}

\item{coupon}{Annnual coupon amount}

\item{conv}{Daycount convention (one of \code{"30/360"}, \code{"Act/Act"},
or \code{"Act/365"})}

\item{freq}{Coupon frequency (number of periods per year)}

\item{clean_px}{The clean price of the bond}
}
\value{
Numeric value of the yield/IRR for the given cashflows and price
}
\description{
Used as the backbone for all yield calculations. This function uses
\code{discount_cfs_single} to generate the cashflows' yearfracs
from the settlement date. The \code{ytm} function from the NMOF manual is
used to calculate the yield, with a 5\% initial guess.
}
\examples{
cfs <- create_cashflows("2045-02-15", 2.5, "2015-05-07", "Act/Act", 2, 100, "ABCD")
yield_cfs_single(cfs, "2015-05-07", 4, "30/360", 2, 101)
}
\seealso{
Other pricing functions: \code{\link{acc_int}};
  \code{\link{calc_clean_px}},
  \code{\link{calc_clean_px.bond}},
  \code{\link{calc_clean_px.default}};
  \code{\link{calc_dirty_px}},
  \code{\link{calc_dirty_px.bond}},
  \code{\link{calc_dirty_px.default}};
  \code{\link{calc_dv01}}, \code{\link{calc_dv01.bond}},
  \code{\link{calc_dv01.default}};
  \code{\link{calc_yield}}, \code{\link{calc_yield.bond}},
  \code{\link{calc_yield.default}};
  \code{\link{discount_cfs_single}};
  \code{\link{dv01_cfs_single}}; \code{\link{ytm}}
}

