% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{ytm}
\alias{ytm}
\title{Yield or IRR for a Single Set of Cashflows with the Outflow}
\usage{
ytm(cf, times, y0 = 0.05, tol = 1e-05, maxit = 1000L)
}
\arguments{
\item{cf}{Vector of cashflow amounts (first cashflow is an outflow)}

\item{times}{Vector of yearfracs (times to cashflow, first should be 0)}

\item{y0}{Initial yield guess (5\% default)}

\item{tol}{Tolerance in the yield calculation}

\item{maxit}{Maximum number of iterations}
}
\value{
Numeric value of the yield/IRR for the given cashflow
amounts and yearfracs in decimal form (not percentage) (e.g. .05)
}
\description{
Used in \code{yield_cfs_single} to do the actual yield calc.
The \code{ytm} function from the NMOF manual is used to calculate
the yield, with a 5\% initial guess. It assumes that the first
cashflow in the \code{cf} vector is negative as it is an outflow equal
to the present value of the cashflows. The corresponding element
in the \code{times} vector should be 0.
}
\examples{
cf <- c(-101, 3, 3, 3, 103)
times <- c(0, 1, 2, 3, 4)
ytm(cf, times)
}
\seealso{
Other pricing functions: \code{\link{acc_int}};
  \code{\link{calc_clean_px}},
  \code{\link{calc_clean_px.bond}},
  \code{\link{calc_clean_px.default}};
  \code{\link{calc_dirty_px}},
  \code{\link{calc_dirty_px.bond}},
  \code{\link{calc_dirty_px.default}};
  \code{\link{calc_risk}}, \code{\link{calc_risk.bond}},
  \code{\link{calc_risk.default}};
  \code{\link{calc_yield}}, \code{\link{calc_yield.bond}},
  \code{\link{calc_yield.default}};
  \code{\link{discount_cfs_single}};
  \code{\link{risk_cfs_single}};
  \code{\link{yield_cfs_single}}
}

