% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonds.R
\name{acc_int}
\alias{acc_int}
\title{Calculate Accrued Interest}
\usage{
acc_int(accrual_start, settle, next_coupon, coupon, conv = "30/360",
  freq = 2)
}
\arguments{
\item{accrual_start}{Accrual start date}

\item{settle}{Settlement Date}

\item{next_coupon}{Next coupon date}

\item{coupon}{Annual coupon amount}

\item{conv}{Daycount convention (one of \code{"30/360"}, \code{"Act/Act"},
or \code{"Act/365"})}

\item{freq}{Coupon frequency (number of periods per year)}
}
\value{
Numeric value of the accrued interest for a par amount of 100
}
\description{
Bonds typically trade based on a "clean price," which does not include
accrued interest. To calculate yields etc. "dirty price" must first be
calculated by including accrued interest. This function assumes that the
dates passed in confrom to the convention that \code{settle} is between
\code{accrual_start} and \code{next_coupon}.
}
\examples{
acc_int("2015-01-15", "2015-03-07", "2015-07-15", "30/360", 2)
}
\seealso{
Other pricing functions: \code{\link{calc_clean_px}},
  \code{\link{calc_clean_px.bond}},
  \code{\link{calc_clean_px.default}};
  \code{\link{calc_dirty_px}},
  \code{\link{calc_dirty_px.bond}},
  \code{\link{calc_dirty_px.default}};
  \code{\link{calc_dv01}}, \code{\link{calc_dv01.bond}},
  \code{\link{calc_dv01.default}}, \code{\link{calc_risk}},
  \code{\link{calc_risk.bond}},
  \code{\link{calc_risk.default}};
  \code{\link{calc_yield}}, \code{\link{calc_yield.bond}},
  \code{\link{calc_yield.default}};
  \code{\link{discount_cfs_single}};
  \code{\link{risk_cfs_single}};
  \code{\link{yield_cfs_single}}; \code{\link{ytm}}
}

