% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\name{load_config}
\alias{load_config}
\title{Load Configurations from a File}
\usage{
load_config(filename)
}
\arguments{
\item{filename}{Path and filename to be loaded}
}
\value{
Environment containing all the assigned variables
}
\description{
Given a .txt filename, this will parse out key value pairs separated
by "=" and assign the value to the key in a new environment.
The file should end in a new line character to avoid any warnings. Users
can access the variables with the base R command \code{get()}
}

